+++ src/test/java/acme/testing/student/activity/StudentActivityShowTest.java
@@ -1,98 +1,98 @@
-
-package acme.testing.student.activity;
-
-import java.util.Collection;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvFileSource;
-import org.springframework.beans.factory.annotation.Autowired;
-
-import acme.entities.Activity;
-import acme.testing.TestHarness;
-
-public class StudentActivityShowTest extends TestHarness {
-
-	@Autowired
-	protected StudentActivityTestRepository repository;
-
-
-	@ParameterizedTest
-	@CsvFileSource(resources = "/student/activity/show-positive.csv", encoding = "utf-8", numLinesToSkip = 1)
-	public void test100Positive(final int enrolmentRecordIndex, final String reference, final int activityRecordIndex, final String title, final String abstract$, final String indication, final String startDate, final String endDate, final String link) {
-		// HINT: this test signs in as an student, lists his or her enrolments, list its activities and checks that it's as expected.
-
-		super.signIn("student1", "student1");
-
-		super.clickOnMenu("Student", "My enrolments");
-		super.checkListingExists();
-		super.sortListing(0, "asc");
-		super.clickOnListingRecord(enrolmentRecordIndex);
-		super.clickOnButton("Activities");
-		super.checkListingExists();
-		super.clickOnListingRecord(activityRecordIndex);
-		super.checkFormExists();
-
-		super.checkInputBoxHasValue("title", title);
-		super.checkInputBoxHasValue("abstract$", abstract$);
-		super.checkInputBoxHasValue("indication", indication);
-		super.checkInputBoxHasValue("startDate", startDate);
-		super.checkInputBoxHasValue("endDate", endDate);
-		super.checkInputBoxHasValue("link", link);
-
-		super.signOut();
-	}
-
-	@Test
-	public void test200Negative() {
-		// HINT: there's no negative test case for this listing, since it doesn't involve filling in any forms.
-	}
-
-	@Test
-	public void test300Hacking() {
-
-		Collection<Activity> activities;
-		String param;
-
-		super.signIn("student1", "student1");
-		activities = this.repository.findManyActivitiesByStudentUsername("student2");
-		for (final Activity activity : activities)
-			if (!activity.getEnrolment().isDraftMode()) {
-				param = String.format("id=%d", activity.getEnrolment().getId());
-
-				super.checkLinkExists("Sign in");
-				super.request("/student/activity/show", param);
-				super.checkPanicExists();
-
-				super.signIn("administrator", "administrator");
-				super.request("/student/activity/show", param);
-				super.checkPanicExists();
-				super.signOut();
-
-				super.signIn("student2", "student2");
-				super.request("/student/activity/show", param);
-				super.checkPanicExists();
-				super.signOut();
-
-				super.signIn("auditor1", "auditor1");
-				super.request("/student/activity/show", param);
-				super.checkPanicExists();
-				super.signOut();
-
-				super.signIn("assistant1", "assistant1");
-				super.request("/student/activity/show", param);
-				super.checkPanicExists();
-				super.signOut();
-
-				super.signIn("company1", "company1");
-				super.request("/student/activity/show", param);
-				super.checkPanicExists();
-				super.signOut();
-
-				super.signIn("lecturer1", "lecturer1");
-				super.request("/student/activity/show", param);
-				super.checkPanicExists();
-				super.signOut();
-			}
-	}
-}
+//
+//package acme.testing.student.activity;
+//
+//import java.util.Collection;
+//
+//import org.junit.jupiter.api.Test;
+//import org.junit.jupiter.params.ParameterizedTest;
+//import org.junit.jupiter.params.provider.CsvFileSource;
+//import org.springframework.beans.factory.annotation.Autowired;
+//
+//import acme.entities.Activity;
+//import acme.testing.TestHarness;
+//
+//public class StudentActivityShowTest extends TestHarness {
+//
+//	@Autowired
+//	protected StudentActivityTestRepository repository;
+//
+//
+//	@ParameterizedTest
+//	@CsvFileSource(resources = "/student/activity/show-positive.csv", encoding = "utf-8", numLinesToSkip = 1)
+//	public void test100Positive(final int enrolmentRecordIndex, final String reference, final int activityRecordIndex, final String title, final String abstract$, final String indication, final String startDate, final String endDate, final String link) {
+//		// HINT: this test signs in as an student, lists his or her enrolments, list its activities and checks that it's as expected.
+//
+//		super.signIn("student1", "student1");
+//
+//		super.clickOnMenu("Student", "My enrolments");
+//		super.checkListingExists();
+//		super.sortListing(0, "asc");
+//		super.clickOnListingRecord(enrolmentRecordIndex);
+//		super.clickOnButton("Activities");
+//		super.checkListingExists();
+//		super.clickOnListingRecord(activityRecordIndex);
+//		super.checkFormExists();
+//
+//		super.checkInputBoxHasValue("title", title);
+//		super.checkInputBoxHasValue("abstract$", abstract$);
+//		super.checkInputBoxHasValue("indication", indication);
+//		super.checkInputBoxHasValue("startDate", startDate);
+//		super.checkInputBoxHasValue("endDate", endDate);
+//		super.checkInputBoxHasValue("link", link);
+//
+//		super.signOut();
+//	}
+//
+//	@Test
+//	public void test200Negative() {
+//		// HINT: there's no negative test case for this listing, since it doesn't involve filling in any forms.
+//	}
+//
+//	@Test
+//	public void test300Hacking() {
+//
+//		Collection<Activity> activities;
+//		String param;
+//
+//		super.signIn("student1", "student1");
+//		activities = this.repository.findManyActivitiesByStudentUsername("student2");
+//		for (final Activity activity : activities)
+//			if (!activity.getEnrolment().isDraftMode()) {
+//				param = String.format("id=%d", activity.getEnrolment().getId());
+//
+//				super.checkLinkExists("Sign in");
+//				super.request("/student/activity/show", param);
+//				super.checkPanicExists();
+//
+//				super.signIn("administrator", "administrator");
+//				super.request("/student/activity/show", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//
+//				super.signIn("student2", "student2");
+//				super.request("/student/activity/show", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//
+//				super.signIn("auditor1", "auditor1");
+//				super.request("/student/activity/show", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//
+//				super.signIn("assistant1", "assistant1");
+//				super.request("/student/activity/show", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//
+//				super.signIn("company1", "company1");
+//				super.request("/student/activity/show", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//
+//				super.signIn("lecturer1", "lecturer1");
+//				super.request("/student/activity/show", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//			}
+//	}
+//}
