+++ src/test/java/acme/testing/student/activity/StudentActivityUpdateTest.java
@@ -1,0 +1,140 @@
+
+package acme.testing.student.activity;
+
+import java.util.Collection;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvFileSource;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import acme.entities.Activity;
+import acme.testing.TestHarness;
+
+public class StudentActivityUpdateTest extends TestHarness {
+
+	@Autowired
+	protected StudentActivityTestRepository repository;
+
+
+	@ParameterizedTest
+	@CsvFileSource(resources = "/student/Activity/update-positive.csv", encoding = "utf-8", numLinesToSkip = 1)
+	public void test100Positive(final int enrolmentRecordIndex, final int activityRecordIndex, final String title, final String abstract$, final String indication, final String startDate, final String endDate, final String link) {
+		// HINT: this test logs in as a student, lists his or her enrolments, selects one of them, updates an activity, and then checks that the update has actually been performed.
+
+		super.signIn("student1", "student1");
+
+		super.clickOnMenu("Student", "My enrolments");
+		super.checkListingExists();
+		super.sortListing(0, "asc");
+
+		super.clickOnListingRecord(enrolmentRecordIndex);
+		super.checkFormExists();
+		super.clickOnButton("Activities");
+		super.checkListingExists();
+		super.clickOnListingRecord(activityRecordIndex);
+
+		super.fillInputBoxIn("title", title);
+		super.fillInputBoxIn("abstract$", abstract$);
+		super.fillInputBoxIn("indication", indication);
+		super.fillInputBoxIn("startDate", startDate);
+		super.fillInputBoxIn("endDate", endDate);
+		super.fillInputBoxIn("link", link);
+
+		super.clickOnSubmit("Update");
+
+		super.checkListingExists();
+		super.sortListing(0, "asc");
+
+		super.clickOnListingRecord(activityRecordIndex);
+		super.checkFormExists();
+
+		super.checkInputBoxHasValue("title", title);
+		super.checkInputBoxHasValue("abstract$", abstract$);
+		super.checkInputBoxHasValue("indication", indication);
+		super.checkInputBoxHasValue("startDate", startDate);
+		super.checkInputBoxHasValue("endDate", endDate);
+		super.checkInputBoxHasValue("link", link);
+
+		super.signOut();
+	}
+	@ParameterizedTest
+	@CsvFileSource(resources = "/student/activity/update-negative.csv", encoding = "utf-8", numLinesToSkip = 1)
+	public void test200Negative(final int enrolmentRecordIndex, final int activityRecordIndex, final String title, final String abstract$, final String indication, final String startDate, final String endDate, final String link) {
+		// HINT: this test attempts to update an activity with wrong data.
+
+		super.signIn("student1", "student1");
+
+		super.clickOnMenu("Student", "My enrolments");
+		super.checkListingExists();
+		super.sortListing(0, "asc");
+
+		super.clickOnListingRecord(enrolmentRecordIndex);
+		super.checkFormExists();
+		super.clickOnButton("Activities");
+		super.checkListingExists();
+		super.sortListing(0, "asc");
+		super.clickOnListingRecord(activityRecordIndex);
+
+		super.fillInputBoxIn("title", title);
+		super.fillInputBoxIn("abstract$", abstract$);
+		super.fillInputBoxIn("indication", indication);
+		super.fillInputBoxIn("startDate", startDate);
+		super.fillInputBoxIn("endDate", endDate);
+		super.fillInputBoxIn("link", link);
+
+		super.clickOnSubmit("Update");
+
+		super.checkErrorsExist();
+
+		super.signOut();
+	}
+
+	@Test
+	public void test300Hacking() {
+		// HINT: this test tries to update an activity with a role other than "Student", or using an student who is not the owner.
+
+		Collection<Activity> activities;
+		String param;
+
+		activities = this.repository.findManyActivitiesByStudentUsername("student1");
+		for (final Activity activity : activities) {
+			param = String.format("id=%d", activity.getId());
+
+			super.checkLinkExists("Sign in");
+			super.request("/student/activity/update", param);
+			super.checkPanicExists();
+
+			super.signIn("administrator", "administrator");
+			super.request("/student/activity/update", param);
+			super.checkPanicExists();
+			super.signOut();
+
+			super.signIn("student2", "student2");
+			super.request("/student/activity/update", param);
+			super.checkPanicExists();
+			super.signOut();
+
+			super.signIn("assistant1", "assistant1");
+			super.request("/student/activity/update", param);
+			super.checkPanicExists();
+			super.signOut();
+
+			super.signIn("company1", "company1");
+			super.request("/student/activity/update", param);
+			super.checkPanicExists();
+			super.signOut();
+
+			super.signIn("lecturer1", "lecturer1");
+			super.request("/student/activity/update", param);
+			super.checkPanicExists();
+			super.signOut();
+
+			super.signIn("auditor1", "auditor1");
+			super.request("/student/activity/update", param);
+			super.checkPanicExists();
+			super.signOut();
+		}
+	}
+
+}
