+++ src/test/resources/student/activity/update-positive.csv
@@ -1,0 +1,5 @@
+enrolmentRecordIndex,activityRecordIndex,title,abstract$,indication,startDate,endDate,link
+0,0,0título de prueba 1,resumen de prueba 1,HANDS_ON,2024/02/01 17:00,2024/02/01 17:30,http://www.youtube.com
+0,0,0título de prueba 1,resumen de prueba 1,THEORY,2024/02/01 17:00,2024/02/01 17:30,
+0,1,1título de prueba 2,resumen de prueba 2,THEORY,2024/03/15 17:00,2024/03/25 17:30,http://www.youtube.com
+0,1,1título de prueba 2,resumen de prueba 2,HANDS_ON,2024/03/15 09:15,2024/03/15 10:20,
